requirements.txt
Django
django-widget-tweaks
djangorestframework
Pillow
django-ckeditor
mysqlclient
pymysql
gunicorn